WebCore/requirements.txt
anyio==3.5.0
asgiref==3.5.0
autopep8==2.0.1
bcrypt==4.0.1
boto3==1.26.39
botocore==1.29.39
cffi==1.15.1
click==8.0.3
colorama==0.4.4
cryptography==38.0.4
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.73.0
FastAPI-SQLAlchemy==0.2.1
greenlet==2.0.1
h11==0.13.0
httptools==0.3.0
idna==3.3
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2-binary
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==6.0
s3transfer==0.6.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.45
starlette==0.17.1
tomli==2.0.1
typing-extensions==4.0.1
urllib3==1.26.13
uvicorn==0.17.1
uvloop==0.17.0
watchfiles==0.18.1
watchgod==0.7
websockets==10.1
